+++ src/main/scala/tscfg/generators/scala/ScalaGen.scala
@@ -1,6 +1,6 @@
 package tscfg.generators.scala
 
-import tscfg.{ModelBuilder, Namespace, model}
+import tscfg.{ModelBuilder, Namespace, model, util}
 import tscfg.generators._
 import tscfg.model._
 import tscfg.util.escapeString
@@ -36,8 +36,12 @@ class ScalaGen(genOpts: GenOpts) extends Generator(genOpts) {
     val packageStr = s"package ${genOpts.packageName}\n\n"
 
     val definition = (packageStr + res.definition).trim
-    //genResults.copy(code = definition)
-    genResults.copy(code = formatter.format(genOpts.packageName, definition))
+    genResults.copy(code = {
+      if (util.doFormatting)
+        formatter.format(genOpts.packageName, definition)
+      else
+        definition
+    })
   }
 
   private def generate(typ: Type,