+++ src/main/scala/tscfg/util.scala
@@ -36,6 +36,9 @@ object util {
       )
       .replace()
   }
+
+  val doFormatting: Boolean = !sys.env.contains("NO_FORMATTER")
+
   // $COVERAGE-ON$
 
   val TypesafeConfigClassName: String = classOf[com.typesafe.config.Config].getName